Fix hiscore restore, rebalance audio
This commit fixes the bug restoring hiscores where it would zero out
the RAM rather than set it to the correct value, and also rebalances
the audio. Previously, the audio on Bagman was such that the speech
was too low relative to the background music. Also, the music for
all the other titles was too loud.